analysis/requirements.txt
requests
numpy
pillow
opencv-python
scikit-learn
scipy
pandas
python-dotenv
